% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combineSingleJobs.R
\name{combineSingleJobs}
\alias{combineSingleJobs}
\title{Network inference module: Combine individual sub-networks to a global network}
\usage{
combineSingleJobs(networkName, cores, path, output = TRUE)
}
\arguments{
\item{networkName}{Name of network}

\item{cores}{Fixed number of sub-network inference problems}

\item{path}{Project path}

\item{output}{Show progress information. Default: TRUE}
}
\value{
Saves global network and overview statistics. See output for details.
}
\description{
This function integrates individual sub-networks into a global network.
}
\examples{

projectPath = getProjectPath( projectName = "MyFirstNetwork", path = "/home/seifert/regNet/AstrocytomaGrades/" )
combineSingleJobs( networkName = "AS_SignatureTFs", cores = 4, path = projectPath )

}
\seealso{
\code{\link{learnNetwork_ParallelComputation}}, \code{\link{loadNetworkWithFilteringForSignificantPredictors}}
}
