% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeNetworkFlow_NewData.R
\name{computeBasicNetworkFlowMatrixBasedOnGivenCorStat_PatientSpecificRelativeImpacts}
\alias{computeBasicNetworkFlowMatrixBasedOnGivenCorStat_PatientSpecificRelativeImpacts}
\title{Network propagation module: Compute basic initial relative patient-specific impact matrix}
\usage{

  computeBasicNetworkFlowMatrixBasedOnGivenCorStat_PatientSpecificRelativeImpacts(patient,
  data, dataSetName, corStatDataSetName, networkName, pValCutoff,
  localGeneCutoff, path, output = TRUE)
}
\arguments{
\item{patient}{Patient considered for impact matrix computation}

\item{data}{Data set containing gene expression and gene copy number profiles for impact computations}

\item{dataSetName}{Name of the data set}

\item{corStatDataSetName}{Name of the data set from which pre-computed correlation statistics are taken}

\item{networkName}{Name of the network}

\item{pValCutoff}{Cutoff for significant links. Values from zero (most sparse network) to one (full network) are allowed.}

\item{localGeneCutoff}{Cutoff for removement of regulator links from/to genes in close chromosomal proximity. Set this to zero if no links should be removed.}

\item{path}{Project path}

\item{output}{Show progress information. Default: TRUE}
}
\value{
Saves basic initial impact matrix. See output for details.
}
\description{
This function computes a basic initial relative patient-specific impact matrix utilizing a given correlation statistics. This function provides the basis for the function \code{\link{computeNetworkFlowBasedOnGivenCorStat_PatientSpecificRelativeImpacts}}. This function allows to separate the computation of the initial from the final impact matrix. See \code{\link{computeNetworkFlowMatrixBasedOnGivenCorStat_PatientSpecificRelativeImpacts}} and \code{\link{computeNetworkFlowMatrix_PatientSpecificRelativeImpacts}} for more details.
}
\examples{

#Load test and train data set and make test data set compatible to the train data set
testData  = loadGeneExpressionAndCopyNumberDataSet( geneExpressionFile = "TCGA_GBM_ExpressionLevels.txt", geneCopyNumberFile = "TCGA_GBM_CopyNumbers.txt", path = "/home/seifert/regNet/AstrocytomaGrades/Data/" )
trainData = loadGeneExpressionAndCopyNumberDataSet( geneExpressionFile = "AS_SignatureTFs_ExpressionLevels.txt", geneCopyNumberFile = "AS_SignatureTFs_CopyNumbers", path = "/home/seifert/regNet/AstrocytomaGrades/Data/" )
gbmTFData = makeTestDataSetCompatible( testDataSet = testData, trainDataSet = trainData )

#Determine correlation statistics
projectPath = getProjectPath( projectName = "MyFirstNetwork", path = "/home/seifert/regNet/AstrocytomaGrades/" )
predictGeneExpression( data = trainData, dataSetName = "AS_SignatureTFs", networkName = "AS_SignatureTFs", pValCutoff = 0.01, localGeneCutoff = 0, path = projectPath )

#Compute impact matrix for first patient using pre-computed correlation statistics
computeBasicNetworkFlowMatrixBasedOnGivenCorStat_PatientSpecificRelativeImpacts( patient = 1, data = testData, dataSetName = "TCGA_GBM_SignatureTFs", corStatDataSetName = "AS_SignatureTFs", networkName = "AS_SignatureTFs", pValCutoff = 0.01, localGeneCutoff = 0, path = projectPath )

}
\seealso{
\code{\link{computeNetworkFlowMatrixBasedOnGivenCorStat_PatientSpecificRelativeImpacts}}, \code{\link{computeNetworkFlowMatrix_PatientSpecificRelativeImpacts}}
}
