% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeNetworkFlow.R
\name{computeNetworkFlow_CohortSpecificAbsoluteImpacts}
\alias{computeNetworkFlow_CohortSpecificAbsoluteImpacts}
\title{Network propagation module: Compute absolute cohort-specific impact matrix}
\usage{
computeNetworkFlow_CohortSpecificAbsoluteImpacts(dataSetName, networkName,
  pValCutoff, localGeneCutoff, colSumsThreshold, path, output = TRUE,
  patientSpecific = FALSE)
}
\arguments{
\item{dataSetName}{Name of the data set}

\item{networkName}{Name of the network}

\item{pValCutoff}{Cutoff for significant links. Values from zero (most sparse network) to one (full network) are allowed.}

\item{localGeneCutoff}{Cutoff for removement of regulator links from/to genes in close chromosomal proximity. Set this to zero if no links should be removed.}

\item{colSumsThreshold}{Stop iterative computation of impact matrix if improvement is less than the given threshold}

\item{path}{Project path}

\item{output}{Show progress information. Default: TRUE}
}
\value{
Saves final impact matrix. See output for details.
}
\description{
This function computes an absolute cohort-specific impact matrix for a given data set (dataSetName) under consideration of a given network (networkName) using network propagation. The basis of this computation is the pre-computed basic initial cohort-specific impact matrix that only considered significant regulatory links (pValCutoff) that were not incoming from genes in close chromosomal proximity up- and downstream (localGeneCutoff) of each specific gene (\code{\link{computeBasicNetworkFlowMatrix_CohortSpecificAbsoluteImpacts}}). The absolute cohort-specific impact matrix is computed iteratively stopping if the sum of the column sums of the current impact matrix increases less than a given threshold (colSumsThreshold) in comparison to the previous impact matrix. The resulting final impact matrix is saved using pre-defined file naming conventions. The resulting final impact matrix can be used to determine absolute impacts of source on target genes. Absolute impacts allow to identify those genes that have the greatest impact on other genes. Absolute impacts do not distinguish between potential activator and inhibitory contributions along the network paths. This function allows to separate the computation of the initial from the final impact matrix.
}
\examples{

projectPath = getProjectPath( projectName = "MyFirstNetwork", path = "/home/seifert/regNet/AstrocytomaGrades/" )
data = loadGeneExpressionAndCopyNumberDataSet( geneExpressionFile = "AS_SignatureTFs_ExpressionLevels.txt", geneCopyNumberFile = "AS_SignatureTFs_CopyNumbers", path = "/home/seifert/regNet/AstrocytomaGrades/Data/" )
computeNetworkFlow_CohortSpecificAbsoluteImpacts( dataSetName = "AS_SignatureTFs", networkName = "AS_SignatureTFs", pValCutoff = 0.01, localGeneCutoff = 0, colSumsThreshold = 1e-3, path = projectPath )

}
\seealso{
\code{\link{computeBasicNetworkFlowMatrix_CohortSpecificAbsoluteImpacts}}, \code{\link{computeNetworkFlowMatrix_CohortSpecificAbsoluteImpacts}}
}
