% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeNetworkFlow.R
\name{computeNetworkFlow_PatientSpecificRelativeImpacts}
\alias{computeNetworkFlow_PatientSpecificRelativeImpacts}
\title{Network propagation module: Compute relative patient-specific impact matrix}
\usage{
computeNetworkFlow_PatientSpecificRelativeImpacts(patient, dataSetName,
  networkName, pValCutoff, localGeneCutoff, colSumsThreshold, path,
  output = TRUE)
}
\arguments{
\item{patient}{Patient considered for impact matrix computation}

\item{dataSetName}{Name of the data set}

\item{networkName}{Name of the network}

\item{pValCutoff}{Cutoff for significant links. Values from zero (most sparse network) to one (full network) are allowed.}

\item{localGeneCutoff}{Cutoff for removement of regulator links from/to genes in close chromosomal proximity. Set this to zero if no links should be removed.}

\item{colSumsThreshold}{Stop iterative computation of impact matrix if global change of impacts is less than the given threshold}

\item{path}{Project path}

\item{output}{Show progress information. Default: TRUE}
}
\value{
Saves final impact matrix. See output for details.
}
\description{
This function computes a relative patient-specific impact matrix for a patient (patient) in a given data set (dataSetName) under consideration of a given network (networkName) using network propagation. The basis of this computation is the pre-computed basic initial patient-specific impact matrix that only considered significant regulatory links (pValCutoff) that were not incoming from genes in close chromosomal proximity up- and downstream (localGeneCutoff) of each specific gene (\code{\link{computeBasicNetworkFlowMatrix_PatientSpecificRelativeImpacts}}). The relative patient-specific impact matrix is computed iteratively stopping if the sum of the column sums of the current impact matrix and the previous impact matrix differ less than a given threshold (colSumsThreshold). The resulting final impact matrix is saved using pre-defined file naming conventions. The resulting final impact matrix can be used to determine relative impacts of source on target genes. Relative impacts account for potential activator and inhibitory contributions along the network paths. This function allows to separate the computation of the initial from the final impact matrix.
}
\examples{

projectPath = getProjectPath( projectName = "MyFirstNetwork", path = "/home/seifert/regNet/AstrocytomaGrades/" )
data = loadGeneExpressionAndCopyNumberDataSet( geneExpressionFile = "AS_SignatureTFs_ExpressionLevels.txt", geneCopyNumberFile = "AS_SignatureTFs_CopyNumbers", path = "/home/seifert/regNet/AstrocytomaGrades/Data/" )
computeNetworkFlow_PatientSpecificRelativeImpacts( patient = 1, dataSetName = "AS_SignatureTFs", networkName = "AS_SignatureTFs", pValCutoff = 0.01, localGeneCutoff = 0, colSumsThreshold = 1e-3, path = projectPath )

}
\seealso{
\code{\link{computeBasicNetworkFlowMatrix_PatientSpecificAbsoluteImpacts}}, \code{\link{computeNetworkFlowMatrix_PatientSpecificAbsoluteImpacts}}
}
