% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeNetworkFlow.R
\name{computeNetworkFlowMatrix_PatientSpecificRelativeImpacts}
\alias{computeNetworkFlowMatrix_PatientSpecificRelativeImpacts}
\title{Network propagation module: Compute relative patient-specific impact matrix}
\usage{
computeNetworkFlowMatrix_PatientSpecificRelativeImpacts(patient, data,
  dataSetName, networkName, pValCutoff, localGeneCutoff, colSumsThreshold, path,
  output = TRUE)
}
\arguments{
\item{patient}{Patient considered for impact matrix computation}

\item{data}{Data set containing gene expression and gene copy number profiles for impact computations}

\item{dataSetName}{Name of the data set}

\item{networkName}{Name of the network}

\item{pValCutoff}{Cutoff for significant links. Values from zero (most sparse network) to one (full network) are allowed.}

\item{localGeneCutoff}{Cutoff for removement of regulator links from/to genes in close chromosomal proximity. Set this to zero if no links should be removed.}

\item{colSumsThreshold}{Stop iterative computation of impact matrix if global change of impacts is less than the given threshold}

\item{path}{Project path}

\item{output}{Show progress information. Default: TRUE}
}
\value{
Saves basic initial and resulting final impact matrices. See output for details.
}
\description{
This function computes a relative patient-specific impact matrix for a patient (patient) part of a given data set (data) under consideration of a given network (networkName) using network propagation. Only significant regulatory links (pValCutoff) that are not incoming from genes in close chromosomal proximity up- and downstream (localGeneCutoff) of each specific gene are included in the computations. First, a basic initial patient-specific matrix is computed. Next, this matrix is used to iteratively compute the relative patient-specific impact matrix stopping if the sum of the column sums of the current impact matrix and the previous impact matrix differ less than a given threshold (colSumsThreshold). The basic inital impact matrix and the resulting final impact matrix are saved using pre-defined file naming conventions. The resulting final impact matrix can be used to determine relative impacts of source on target genes. Relative impacts account for potential activator and inhibitory contributions along the network paths.
}
\examples{

projectPath = getProjectPath( projectName = "MyFirstNetwork", path = "/home/seifert/regNet/AstrocytomaGrades/" )
data = loadGeneExpressionAndCopyNumberDataSet( geneExpressionFile = "AS_SignatureTFs_ExpressionLevels.txt", geneCopyNumberFile = "AS_SignatureTFs_CopyNumbers", path = "/home/seifert/regNet/AstrocytomaGrades/Data/" )

#Compute impact matrix for first patient of the given data set
computeNetworkFlowMatrix_PatientSpecificRelativeImpacts( patient = 1, data = data, dataSetName = "AS_SignatureTFs", networkName = "AS_SignatureTFs", pValCutoff = 0.01, localGeneCutoff = 0, colSumsThreshold = 1e-3, path = projectPath )

}
\seealso{
\code{\link{computeBasicNetworkFlowMatrix_PatientSpecificAbsoluteImpacts}}, \code{\link{computeNetworkFlow_PatientSpecificAbsoluteImpacts}}, \code{\link{loadNetworkFlowMatrix_PatientSpecificAbsoluteImpacts}}, \code{\link{getAverageImpacts_PatientSpecificAbsoluteImpacts}}, \code{\link{getImpacts_PatientSpecificAbsoluteImpacts}}, \code{\link{getRawImpacts_PatientSpecificAbsoluteImpacts}}, \code{\link{computeNetworkFlowMatrix_CohortSpecificAbsoluteImpacts}}, \code{\link{computeNetworkFlowMatrix_PatientSpecificRelativeImpacts}}, ...
}
