% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeNetworkFlow_NewData.R
\name{computeNetworkFlowMatrixBasedOnGivenCorStat_PatientSpecificRelativeImpacts}
\alias{computeNetworkFlowMatrixBasedOnGivenCorStat_PatientSpecificRelativeImpacts}
\title{Network propagation module: Compute relative patient-specific impact matrix}
\usage{

  computeNetworkFlowMatrixBasedOnGivenCorStat_PatientSpecificRelativeImpacts(patient,
  data, dataSetName, corStatDataSetName, networkName, pValCutoff,
  localGeneCutoff, colSumsThreshold, path, output = TRUE)
}
\arguments{
\item{patient}{Patient considered for impact matrix computation}

\item{data}{Data set containing gene expression and gene copy number profiles for impact computations}

\item{dataSetName}{Name of the data set}

\item{corStatDataSetName}{Name of the data set from which pre-computed correlation statistics are taken}

\item{networkName}{Name of the network}

\item{pValCutoff}{Cutoff for significant links. Values from zero (most sparse network) to one (full network) are allowed.}

\item{localGeneCutoff}{Cutoff for removement of regulator links from/to genes in close chromosomal proximity. Set this to zero if no links should be removed.}

\item{colSumsThreshold}{Stop iterative computation if the impacts of the current and the previous impact matrix differ less than the given threshold}

\item{path}{Project path}

\item{output}{Show progress information. Default: TRUE}
}
\value{
Saves basic initial and resulting final impact matrices. See output for details.
}
\description{
This function computes a relative patient-specific impact matrix utilizing a given correlation statistics. This enables to analyze a patient (patient) in a new data set (data, dataSetName) under consideration of the predictive power of another data set (corStatDataSetName). This supports the analysis of data sets with few patients where the sample size is too small to obtain a robust estimate of the predictive power of individual genes (correlation statistics). See \code{\link{computeNetworkFlowMatrix_PatientSpecificRelativeImpacts}} for more details. The correlation statistics can be pre-computed using the function \code{\link{predictGeneExpression}}.
}
\examples{

#Load test and train data set and make test data set compatible to the train data set
testData  = loadGeneExpressionAndCopyNumberDataSet( geneExpressionFile = "TCGA_GBM_ExpressionLevels.txt", geneCopyNumberFile = "TCGA_GBM_CopyNumbers.txt", path = "/home/seifert/regNet/AstrocytomaGrades/Data/" )
trainData = loadGeneExpressionAndCopyNumberDataSet( geneExpressionFile = "AS_SignatureTFs_ExpressionLevels.txt", geneCopyNumberFile = "AS_SignatureTFs_CopyNumbers", path = "/home/seifert/regNet/AstrocytomaGrades/Data/" )
gbmTFData = makeTestDataSetCompatible( testDataSet = testData, trainDataSet = trainData )

#Determine correlation statistics
projectPath = getProjectPath( projectName = "MyFirstNetwork", path = "/home/seifert/regNet/AstrocytomaGrades/" )
predictGeneExpression( data = trainData, dataSetName = "AS_SignatureTFs", networkName = "AS_SignatureTFs", pValCutoff = 0.01, localGeneCutoff = 0, path = projectPath )

#Compute impact matrix for first patient using pre-computed correlation statistics
computeNetworkFlowMatrixBasedOnGivenCorStat_PatientSpecificRelativeImpacts( patient = 1, data = testData, dataSetName = "TCGA_GBM_SignatureTFs", corStatDataSetName = "AS_SignatureTFs", networkName = "AS_SignatureTFs", pValCutoff = 0.01, localGeneCutoff = 0, colSumsThreshold = 1e-3, path = projectPath )

}
\seealso{
\code{\link{computeNetworkFlowMatrix_PatientSpecificRelativeImpacts}}
}
