% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createBasicFolderStructure.R
\name{createBasicFolderStructure}
\alias{createBasicFolderStructure}
\title{Data loader module: Create basic folder structure}
\usage{
createBasicFolderStructure(projectName, path, output = TRUE)
}
\arguments{
\item{projectName}{Name of project folder}

\item{path}{Directory where the project folder is created}

\item{output}{Show progress information. Default: TRUE}
}
\value{
String containing the project path.
}
\description{
This function creates the standardized folder structure used by regNet for network inference, prediction, and network propagation.
}
\examples{
createBasicFolderStructure( projectName = "MyFirstNetwork", path = "/home/seifert/regNet/AstrocytomaGrades/" )

}
\seealso{
\code{\link{getProjectPath}}
}
