% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/determineRandomNetwork.R
\name{determineRandomNetworkWithFilteringForSignificantPredictors}
\alias{determineRandomNetworkWithFilteringForSignificantPredictors}
\title{Network inference module: Create random network instance}
\usage{
determineRandomNetworkWithFilteringForSignificantPredictors(networkName,
  pValCutoff, randomNetworkInstance, path, output = TRUE)
}
\arguments{
\item{networkName}{Name of the network}

\item{pValCutoff}{Cutoff for significant links. Values from zero (most sparse network) to one (full network) are allowed.}

\item{randomNetworkInstance}{Random network instance number}

\item{path}{Project path}

\item{output}{Show progress information. Default: TRUE}
}
\value{
Saves random network instance. See output for details.
}
\description{
This function loads a specified network (networkName), filters this network for non-significant regulatory links (pValCutoff), and finally creates a random degree-preserving instance of the network. The resulting random network is saved using a pre-defined file-naming convention including the random network instance number (randomNetworkInstance) to uniquely label the random network for later usage. The random network is compatible with all other regNet functions and can be used to generate baseline results for network predictions and network propagation.
}
\examples{

projectPath = getProjectPath( projectName = "MyFirstNetwork", path = "/home/seifert/regNet/AstrocytomaGrades/" )
determineRandomNetworkWithFilteringForSignificantPredictors( networkName = "AS_SignatureTFs", pValCutoff = 0.01, randomNetworkInstance = 1, path = projectPath )

}
\seealso{
\code{\link{loadNetworkWithFilteringForSignificantPredictors}}, \code{\link{predictGeneExpression}}, ...
}
