% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeImpacts.R
\name{getAverageImpactsBasedOnGivenCorStat_CohortSpecificAbsoluteImpacts}
\alias{getAverageImpactsBasedOnGivenCorStat_CohortSpecificAbsoluteImpacts}
\title{Network propagation module: Determine average absolute cohort-specific impacts}
\usage{
getAverageImpactsBasedOnGivenCorStat_CohortSpecificAbsoluteImpacts(sourceGenes,
  targetGenes, dataSetName, corStatDataSetName, networkName, pValCutoff,
  localGeneCutoff, colSumsThreshold, path, outputFile, output = TRUE)
}
\arguments{
\item{sourceGenes}{Source genes for which average impacts on target genes are determined}

\item{targetGenes}{Target genes}

\item{dataSetName}{Name of the data set}

\item{corStatDataSetName}{Name of the data set from which pre-computed correlation statistics were taken}

\item{networkName}{Name of the network}

\item{pValCutoff}{Cutoff for significant links used for pre-computed impact matrix}

\item{localGeneCutoff}{Cutoff for removement of regulator links from/to genes in close chromosomal proximity used for pre-computed impact matrix}

\item{colSumsThreshold}{Stop criterion used for pre-computed impact matrix}

\item{path}{Project path}

\item{outputFile}{Name of output file}

\item{output}{Show progress information. Default: TRUE}
}
\value{
Returns and saves average impacts. See output for details.
}
\description{
This function determines average absolute cohort-specific impacts for each given source gene (sourceGenes) on all given target genes (targetGenes) utilizing a pre-computed cohort-specific impact matrix that was computed using the predictive power of a given data set (corStatDataSetName). The average impacts are saved to a user-defined file (outputFile) in the pre-defined regNet folder structure. In addition, the average impacts are also directly returned as matrix. Average absolute impacts allow to identify those source genes that have the greatest average impact on target genes.
}
\examples{

projectPath = getProjectPath( projectName = "MyFirstNetwork", path = "/home/seifert/regNet/AstrocytomaGrades/" )
testData  = loadGeneExpressionAndCopyNumberDataSet( geneExpressionFile = "TCGA_GBM_ExpressionLevels.txt", geneCopyNumberFile = "TCGA_GBM_CopyNumbers.txt", path = "/home/seifert/regNet/AstrocytomaGrades/Data/" )
trainData = loadGeneExpressionAndCopyNumberDataSet( geneExpressionFile = "AS_SignatureTFs_ExpressionLevels.txt", geneCopyNumberFile = "AS_SignatureTFs_CopyNumbers", path = "/home/seifert/regNet/AstrocytomaGrades/Data/" )
data = makeTestDataSetCompatible( testDataSet = testData, trainDataSet = trainData )
getAverageImpactsBasedOnGivenCorStat_CohortSpecificAbsoluteImpacts( sourceGenes = c( "TBR1", "CCNL2" ), targetGenes = data$genes, dataSetName = "TCGA_GBM_SignatureTFs", corStatDataSetName = "AS_SignatureTFs", networkName = "AS_SignatureTFs", pValCutoff = 0.01, localGeneCutoff = 0, colSumsThreshold = 1e-3, path = projectPath, outputFile = "AvgImpactsOfSelectedRegulators.txt" )

}
\seealso{
\code{\link{computeBasicNetworkFlowMatrixBasedOnGivenCorStat_CohortSpecificAbsoluteImpacts}}, \code{\link{getAverageImpacts_CohortSpecificAbsoluteImpacts}}
}
