% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learnNetwork_ParallelComputation.R
\name{getComputationRange}
\alias{getComputationRange}
\title{Network inference module: Define sub-network inference problems}
\usage{
getComputationRange(cores, nrResponses)
}
\arguments{
\item{cores}{Fixed number of sub-network inference problems}

\item{nrResponses}{Total number of genes in a data set}
}
\value{
Array of (nearly) equal-sized intervals of genes
}
\description{
This is an internal function that divides the total number of genes (nrResponses) in a data set into a fixed number (cores) of equal-sized intervals of genes. This is required to separate the global network inference problem into independent sub-network inference problems.
}
\examples{

range = getComputationRange( cores = 400, nrResponses = 15811 )

}
