% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNetworkConnections.R
\name{getNetworkConnections}
\alias{getNetworkConnections}
\title{Network inference module: Get network connections}
\usage{
getNetworkConnections(networkName, pValCutoff, localGeneCutoff, path,
  output = TRUE)
}
\arguments{
\item{networkName}{Name of the network}

\item{pValCutoff}{Cutoff for significant links. Values from zero (most sparse network) to one (full network) are allowed.}

\item{localGeneCutoff}{Cutoff for removement of regulator links from/to genes in close chromosomal proximity. Set this to zero if no links should be removed.}

\item{path}{Project path}

\item{output}{Show progress information. Default: TRUE}
}
\value{
Saves a network connectivity table. See output for details.
}
\description{
This function determines learned links between genes in the given network (networkName). These links comprise all incoming and outgoing activator and/or inhibitor links of each individual gene. Also direct gene copy number effects are reported. Only links with p-values equal or below the given p-value cutoff (pValCutoff) are considered. For each specific gene, all incoming/outgoing links from/to genes in close chromosomal proximity up- and downstream of the considered gene are removed (localGeneCutoff).
}
\examples{

projectPath = getProjectPath( projectName = "MyFirstNetwork", path = "/home/seifert/regNet/AstrocytomaGrades/" )
getNetworkConnections( networkName = "AS_SignatureTFs", pValCutoff = 0.01, localGeneCutoff = 0, path = projectPath )

}
\seealso{
\code{\link{getNetworkConnectionsForVisualization}}
}
