% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createBasicFolderStructure.R
\name{getProjectPath}
\alias{getProjectPath}
\title{Data loader module: Get basic folder structure path}
\usage{
getProjectPath(projectName, path)
}
\arguments{
\item{projectName}{Name of project folder}

\item{path}{Directory where the project folder is created}
}
\value{
String containing the project path.
}
\description{
This function returns the project path.
}
\examples{
getProjectPath( projectName = "MyFirstNetwork", path = "/home/seifert/regNet/AstrocytomaGrades/" )

}
\seealso{
\code{\link{createBasicFolderStructure}}
}
