% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeImpacts.R
\name{getRawImpacts_CohortSpecificAbsoluteImpacts}
\alias{getRawImpacts_CohortSpecificAbsoluteImpacts}
\title{Network propagation module: Determine raw absolute cohort-specific impacts}
\usage{
getRawImpacts_CohortSpecificAbsoluteImpacts(sourceGenes, targetGenes,
  dataSetName, networkName, pValCutoff, localGeneCutoff, colSumsThreshold, path,
  outputFile, output = TRUE)
}
\arguments{
\item{sourceGenes}{Source genes for which raw impacts on target genes are determined}

\item{targetGenes}{Target genes}

\item{dataSetName}{Name of the data set}

\item{networkName}{Name of the network}

\item{pValCutoff}{Cutoff for significant links used for pre-computed impact matrix}

\item{localGeneCutoff}{Cutoff for removement of regulator links from/to genes in close chromosomal proximity used for pre-computed impact matrix}

\item{colSumsThreshold}{Stop criterion used for pre-computed impact matrix}

\item{path}{Project path}

\item{outputFile}{Name of output file}

\item{output}{Show progress information. Default: TRUE}
}
\value{
Returns and saves raw impacts. See output for details.
}
\description{
This function determines raw absolute cohort-specific impacts for each given source gene (sourceGenes) on each given target gene (targetGenes) utilizing a pre-computed cohort-specific impact matrix. Raw impacts represent the entries of the impact matrix without furhter standardizing each column sum of the impact matrix to one. The raw impacts are saved to a user-defined file (outputFile) in the pre-defined regNet folder structure. In addition, the raw impacts are also directly returned as matrix. Raw absolute impacts allow to identify those source genes that have the greatest impact on target genes.
}
\examples{

projectPath = getProjectPath( projectName = "MyFirstNetwork", path = "/home/seifert/regNet/AstrocytomaGrades/" )
data = loadGeneExpressionAndCopyNumberDataSet( geneExpressionFile = "AS_SignatureTFs_ExpressionLevels.txt", geneCopyNumberFile = "AS_SignatureTFs_CopyNumbers", path = "/home/seifert/regNet/AstrocytomaGrades/Data/" )
getRawImpacts_CohortSpecificAbsoluteImpacts( sourceGenes = c( "TBR1", "CCNL2" ), targetGenes = data$genes dataSetName = "AS_SignatureTFs", networkName = "AS_SignatureTFs", pValCutoff = 0.01, localGeneCutoff = 0, colSumsThreshold = 1e-3, path = projectPath, outputFile = "AvgImpactsOfSelectedRegulators.txt" )

}
\seealso{
\code{\link{computeNetworkFlowMatrix_CohortSpecificAbsoluteImpacts}}, \code{\link{getRawImpactsBasedOnGivenCorStat_CohortSpecificAbsoluteImpacts}}
}
