% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadDataSet.R
\name{loadDataSet}
\alias{loadDataSet}
\title{Data loader module: Load data set}
\usage{
loadDataSet(file, path)
}
\arguments{
\item{file}{Name of data file}

\item{path}{Data file path}
}
\value{
List containing measurements (data), genes (genes), chromosomes (chr), and location information (loc).
}
\description{
This function loads a data set (e.g. gene expression, gene copy number) from a given file. The data set must contain the following columns: Gene, Chromosome, Loctcation, Measurements_1, ..., Measurements_N. Gene column: unique gene names. Chromosome column: chromosome entry of genes. Location column: position entry of each gene. Measurement column: gene-specific measurements. Columns in the data set must be separated by tabulators. Genes in the data set must be sorted according to their chromosomal locations.
}
\examples{

data = loadDataSet( loadFile = "AS_SignatureTFs_ExpressionLevels.txt", loadPath = "/home/seifert/regNet/AstrocytomaGrades/Data/" )

}
\seealso{
\code{\link{loadGeneExpressionAndCopyNumberDataSet}}
}
