% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadDataSet.R
\name{loadGeneExpressionAndCopyNumberDataSet}
\alias{loadGeneExpressionAndCopyNumberDataSet}
\title{Data loader module: Load data set that can used by regNet for network inference, prediction, or network propagation}
\usage{
loadGeneExpressionAndCopyNumberDataSet(geneExpressionFile, geneCopyNumberFile,
  path)
}
\arguments{
\item{geneExpressionFile}{Gene expression data file}

\item{path}{Path to data files}

\item{geneExpressionFile}{Gene copy number data file}
}
\value{
List containing gene copy number profiles (U), gene expression profiles(Y), genes (genes), chromosomes (chr), and location information (loc).
}
\description{
This function loads gene expression and corresponding gene copy number data from the given files. The obtained data set can be used by regNet for network learning, prediction, or propagation. See \code{\link{loadDataSet}} for information on how to structure both data files.
}
\examples{

data = loadGeneExpressionAndCopyNumberDataSet( geneExpressionFile = "AS_SignatureTFs_ExpressionLevels.txt", geneCopyNumberFile = "AS_SignatureTFs_CopyNumbers", path = "/home/seifert/regNet/AstrocytomaGrades/Data/" )

}
\seealso{
\code{\link{loadDataSet}} which is called by this function
}
