% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadNetwork.R
\name{loadNetworkWithFilteringForSignificantPredictors}
\alias{loadNetworkWithFilteringForSignificantPredictors}
\title{Network inference module: Load network}
\usage{
loadNetworkWithFilteringForSignificantPredictors(networkName, pValCutoff, path,
  output = TRUE)
}
\arguments{
\item{networkName}{Name of the network}

\item{pValCutoff}{Cutoff for significant links. Values from zero (most sparse network) to one (full network) are allowed.}

\item{path}{Project path}

\item{output}{Show progress information. Default: TRUE}
}
\value{
List containing the components of the network. G: network coefficient matrix (rows: response genes, columns: predictors). o: gene-specific offset. pVal: FDR-adjusted p-value matrix belonging to G. responseGenes: Response genes. responseGenesChr: Chromsomes on which the response genes are located. responseGenesLoc: Chromosomal locations of response genes. predictors: Names of preditors.
}
\description{
This function loads a specified network (networkName) and filters for non-significant regulatory links (pValCutoff).
}
\examples{

projectPath = getProjectPath( projectName = "MyFirstNetwork", path = "/home/seifert/regNet/AstrocytomaGrades/" )
loadNetworkWithFilteringForSignificantPredictors( networkName = "AS_SignatureTFs", pValCutoff = 0.01, path = projectPath )

}
\seealso{
\code{\link{learnNetwork_ParallelComputation}}
}
