% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadDataSet.R
\name{makeTestDataSetCompatible}
\alias{makeTestDataSetCompatible}
\title{Data loader module: Transform test data set}
\usage{
makeTestDataSetCompatible(testDataSet, trainDataSet, output = TRUE)
}
\arguments{
\item{testDataSet}{Test data set}

\item{trainDataSet}{Data set used for network inference}

\item{output}{Show progress information. Default: TRUE}
}
\value{
Test data set compatible with training data set
}
\description{
This function transforms a given test data set to make it compatible to a given training data set that was used to learn a network. Additional genes are removed, missing genes are artificially added with measurement values of zero, and the same order of genes is ensured.
}
\examples{

testData  = loadGeneExpressionAndCopyNumberDataSet( geneExpressionFile = "TCGA_GBM_ExpressionLevels.txt", geneCopyNumberFile = "TCGA_GBM_CopyNumbers.txt", path = "/home/seifert/regNet/AstrocytomaGrades/Data/" )
trainData = loadGeneExpressionAndCopyNumberDataSet( geneExpressionFile = "AS_SignatureTFs_ExpressionLevels.txt", geneCopyNumberFile = "AS_SignatureTFs_CopyNumbers", path = "/home/seifert/regNet/AstrocytomaGrades/Data/" )
gbmTFData = makeTestDataSetCompatible( testDataSet = testData, trainDataSet = trainData )

}
\seealso{
\code{\link{loadGeneExpressionAndCopyNumberDataSet}}, \code{\link{testDataSetCompatibility}}
}
