% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.regnet.R
\name{print.cv.regnet}
\alias{print.cv.regnet}
\title{print a cv.glmnet object}
\usage{
\method{print}{cv.regnet}(x, digits = max(3, getOption("digits") - 3), \dots)
}
\arguments{
\item{x}{cv.glmnet object.}

\item{digits}{significant digits in printout.}

\item{...}{other print arguments}
}
\description{
Print a summary of a cv.glmnet object
}
\seealso{
\code{\link{cv.regnet}}
}
