% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.regnet.R
\name{print.regnet}
\alias{print.regnet}
\title{print a regnet object}
\usage{
\method{print}{regnet}(x, digits = max(3, getOption("digits") - 3), \dots)
}
\arguments{
\item{x}{regnet object.}

\item{digits}{significant digits in printout.}

\item{...}{other print arguments}
}
\description{
Print a summary of a regnet object
}
\seealso{
\code{\link{regnet}}
}
