% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regnet.R
\name{regnet.cv.regnet}
\alias{regnet.cv.regnet}
\title{Fit a regnet model from cross-validation results}
\usage{
\method{regnet}{cv.regnet}(X, lamb.1 = NULL, lamb.2 = NULL, ...)
}
\arguments{
\item{X}{a \code{cv.regnet} object returned by \code{\link{cv.regnet}}.}

\item{lamb.1}{the tuning parameter \eqn{\lambda_{1}} that imposes sparsity.}

\item{lamb.2}{the tuning parameter \eqn{\lambda_{2}} that controls smoothness among coefficient profiles. \eqn{\lambda_{2}} is needed for network penalty.}

\item{...}{additional named arguments passed to \code{regnet.default}. These can be used to override settings stored in \code{X}.}
}
\value{
an object of class \code{regnet}.
}
\description{
This method refits \code{\link{regnet}} using the data and model settings
stored in a \code{\link{cv.regnet}} object. The tuning parameter values must
be supplied explicitly through \code{lamb.1} and, for network models,
\code{lamb.2}.
}
\examples{
\donttest{
data(LogisticExample)
X = rgn.logi$X
Y = rgn.logi$Y
cvfit = cv.regnet(X, Y, response="binary", penalty="network", folds=5, r = 4.5)
fit = regnet(cvfit, cvfit$lambda[1,1], cvfit$lambda[1,2])
}

}
\seealso{
\code{\link{cv.regnet}}, \code{\link{regnet}}
}
