% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadDataSet.R
\name{testDataSetCompatibility}
\alias{testDataSetCompatibility}
\title{Data loader module: Test data set compatibility}
\usage{
testDataSetCompatibility(testDataSet, trainDataSet)
}
\arguments{
\item{testDataSet}{Test data set}

\item{trainDataSet}{Data set utilized to learn the network}
}
\value{
Boolean specifying if both data sets are compatible
}
\description{
This function evaluates if the two given data sets loaded by \code{\link{loadGeneExpressionAndCopyNumberDataSet}} are compatible. Compatibility of data sets must be ensured to utilize a learned network for predictions or network flow computations.
}
\examples{

testData  = loadGeneExpressionAndCopyNumberDataSet( geneExpressionFile = "TCGA_GBM_ExpressionLevels.txt", geneCopyNumberFile = "TCGA_GBM_CopyNumbers.txt", path = "/home/seifert/regNet/AstrocytomaGrades/Data/" )
trainData = loadGeneExpressionAndCopyNumberDataSet( geneExpressionFile = "AS_SignatureTFs_ExpressionLevels.txt", geneCopyNumberFile = "AS_SignatureTFs_CopyNumbers", path = "/home/seifert/regNet/AstrocytomaGrades/Data/" )
testDataSetCompatibility( testDataSet = testData, trainDataSet = trainData )

}
\seealso{
\code{\link{loadGeneExpressionAndCopyNumberDataSet}}, \code{\link{makeTestDataSetCompatible}}
}
