% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regnet.R
\docType{data}
\name{regnet}
\alias{regnet}
\title{A simulated dataset for testing logistic regression.}
\format{An object of class \code{list} of length 3.}
\usage{
regnet
}
\description{
The dataset containing the X matrix, binary response Y and the true coefficient vector beta
}
\keyword{datasets}
