% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\docType{data}
\name{rgn}
\alias{rgn}
\alias{rgn.logi}
\alias{rgn.surv}
\title{simulated data for demonstrating the features of regnet}
\format{Each dataset consists of three components: X, Y, and beta; beta is the
true coefficient used to generated Y.}
\usage{
data("LogisticExample")
data("SurvExample")
}
\description{
Simple simulated data for demonstrating the features of regnet.
}
\examples{
data("LogisticExample")
lapply(rgn.logi, class)
}
\keyword{datasets}
