% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\docType{data}
\name{rgn}
\alias{rgn}
\alias{rgn.logi}
\alias{rgn.surv}
\alias{rgn.tcga}
\alias{rgn.htr}
\title{Example datasets for demonstrating the features of regnet}
\format{
"LogisticExample", "SurvExample" and "HeteroExample" are simulated data. Each data includes three main components: X, Y, and beta; beta is a vector of the
true coefficients used to generate Y.

"ContExample" is a subset of the skin cutaneous melanoma data from the Cancer Genome Atlas (TCGA). The response variable Y is
the log-transformed Breslow’s depth. X is a matrix of gene expression data.
}
\usage{
data("LogisticExample")
data("SurvExample")
data("ContExample")
data("HeteroExample")
}
\description{
Example datasets for demonstrating the features of regnet.
}
\examples{
data("LogisticExample")
lapply(rgn.logi, class)
}
\keyword{datasets}
